# Preserve rules priority while expanding shorthands