… test/rule_set/test_declarations.rb