We can piggyback on `Hash` preserving it's order in all supported rubies & drop order. Looks quite ugly, but does the trick.
I assume that it will be faster if we'll properly track and propagate `order` in all places, but I suspect it will be also harder to work with.

And we're little faster at parsing because of `order` is removed:
Before:
```
Executing: bundle exec rake benchmark
Warming up --------------------------------------
 import1.css loading   376.000  i/100ms
 complex.css loading    14.000  i/100ms
Calculating -------------------------------------
 import1.css loading      3.670k (± 3.6%) i/s -     18.424k in   5.027200s
 complex.css loading    142.421  (± 4.2%) i/s -    714.000  in   5.021913s

Loading `import1.css` allocated 372 objects, 26 KiB
Loading `complex.css` allocated 10437 objects, 752 KiB
```
After:
```
Executing: bundle exec rake benchmark
Warming up --------------------------------------
 import1.css loading   394.000  i/100ms
 complex.css loading    14.000  i/100ms
Calculating -------------------------------------
 import1.css loading      3.809k (± 3.7%) i/s -     19.306k in   5.075438s
 complex.css loading    147.464  (± 3.4%) i/s -    742.000  in   5.037139s

Loading `import1.css` allocated 357 objects, 25 KiB
Loading `complex.css` allocated 9536 objects, 680 KiB
```

And little slower at expanding shorthands (tested on github's 500KB css with `parser.each_rule_set { |rule_set, _media_type| rule_set.expand_shorthand! }`, not sure how relevant this file is though)
Before:
```
Executing: bundle exec ruby b.rb
Warming up --------------------------------------
github-bf18d76d7dfd1ad6ed77f31817d6c749.css loading
                         1.000  i/100ms
Calculating -------------------------------------
github-bf18d76d7dfd1ad6ed77f31817d6c749.css loading
                          2.798  (± 0.0%) i/s -     84.000  in  30.089636s
Loading `github.css` and expanding shothands allocated 439154 objects, 29732 KiB
```
After:
```
Executing: bundle exec ruby b.rb
Warming up --------------------------------------
github-bf18d76d7dfd1ad6ed77f31817d6c749.css loading
                         1.000  i/100ms
Calculating -------------------------------------
github-bf18d76d7dfd1ad6ed77f31817d6c749.css loading
                          2.789  (± 0.0%) i/s -     84.000  in  30.161549s
Loading `github.css` and expanding shothands allocated 437074 objects, 30186 KiB
Successfully rebased and updated refs/heads/fixup_expanding_shorthands.
``` 
Fixes #111