@ojab 2021-01-19
So yeah, it's more or less RFC at this point and I will add `Declarations#replace_declaration!` tests, just want to make sure that this approach is acceptable.