@ojab 2021-01-21
Yeah, trying to test forwardable, but looks like it's not trivial here https://github.com/seattlerb/minitest/issues/545