@grosser 2021-01-21
1.8.0

On Thu, Jan 21, 2021 at 11:36 AM Slava Kardakov <notifications@github.com>
wrote:

> Thanks 🎉! Can I haz release?
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/premailer/css_parser/pull/119#issuecomment-764887413>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAACYZ372C227ZMGZHBL4F3S3B62PANCNFSM4WJEACNA>
> .
>